pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "linework"
version = "0.1.0"
description = "Agent-first CLI sketch tool."
readme = "SPEC.md"
requires-python = ">=3.12,<3.13"
license = { text = "Proprietary" }
authors = [
  { name = "Nicolas Villar" },
]
dependencies = [
    "pillow>=12.2.0",
]

[project.scripts]
linework = "linework.cli:main"

[dependency-groups]
dev = [
  "mypy>=1.18.2",
  "pytest>=8.4.2",
  "ruff>=0.13.1",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "UP",
]

[tool.ruff.format]
docstring-code-format = true

[tool.mypy]
python_version = "3.12"
mypy_path = ["src"]
strict = true
warn_unreachable = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-q"]
